requirements.txt
langchain
pypdf
python-box
faiss-cpu
ctransformers
sentence-transformers
